Disable the "resample" filter
This is the magic filter libav uses to automatically convert between
sample rates and is the only filter in ffmpeg which depends on
libavresample.

Given that libavresample is deprecated and may be removed in the future,
it would be useful to have libavfilter not depend on libavresample now
(eg for testing whether a package needs libavresample). Since this
filter is not used by ffmpeg (which uses the "aresample" filter
instead), and the documentation states that this filter should not
be used manually, it should be safe to disable.

Once disabled, we can safely drop some libavresample-dev dependencies.