Add symbols files (again)
Have another attempt at adding symbols files to ffmpeg. Since we've just
had an ABI bump, this is fine for 4.0, but will increase the work needed
to package 4.1 when it comes out :)

All avpriv_ functions are marked as optional because they are not part
of the public ABI. Intra-library dependencies (which need to be tighter)
and handled manually in debian/control.