Only pass --target-os when cross compiling
Since d3d378eb9637 ("Unconditionally pass --arch and --target-os"), the
--target-os option has been passed unconditionally. This breaks the
build on hurd where the value of DEB_HOST_ARCH_OS is not the same as the
uname -s value. Unlike the --arch option, FFmpeg's configure does not
handle the various aliases to allow this to work.

Workaround this for native builds by moving --target-os back into cross
compilation only. Hopefully this can be moved back once upstream is
fixed.

Closes: #904052