Add proper multi-dist build process
Adds the ability to configure and build per-release deb files that
can be used effectively with reprepro to provide compatibility
across distros and multiple architectures.

The package contains a set of static, single-binary ffmpeg and
ffprobe binaries that can be used instead of any system ffmpeg
packages, for whatever reason. ffmpeg 4.0.3 was chosen as the
latest Jellyfin-supported version.

The resulting binary files are located in `/usr/share` rather than
`/usr/bin`, so we don't need to `Conflicts` with system ffmpeg
packages. It does require various shared dependencies, following
as closely as possible (while compatible with all supported
releases) the default configurations, all of which should be
available on all releases and architectures by default (as would
the system ffmpeg).

A Docker-based infrastructure similar to the main Jelyfin package
build infrastructure is included to build the various packages
for the supported releases and architectures. The entrypoint
for users cloning the repo is `build`.

This configuration disables all tests as a number of them fail
arbitrarily for unknown reasons on armhf, but not amd64. This might
not be ideal long-term but provides a temporary solution. Reenabling
tests can be done by removing the `override_dh_auto_test:` line in
the `debian/rules` file.