avfilter/tonemap*: add ICtCp based tonemap method
This adds ICtCp based tonemap to OpenCL, Metal and CUDA based
tonemap filters, and will use the mode as the new default. On GPUs
with lower specs, OpenCL and CUDA implementation will default to
relative luminance mode. The Metal implementation does not have a
performance tradeoff mode and will always default to ICtCp mode.

This also fixed attachment related issue on Metal when output is
10bit by always setting the attachement values.